pyproject.toml
[project]
name = "forest-pulse"
version = "0.1.0"
description = "Aerial imagery tree detection, health scoring, and temporal change monitoring for forest management."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Jordi Catafal" },
]
keywords = ["tree-detection", "forestry", "computer-vision", "aerial-imagery", "health-monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "supervision>=0.25.0",
    "deepforest>=2.0.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "geopandas>=1.0.0",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
    "piexif>=1.1.3",
]

[project.optional-dependencies]
train = [
    "rfdetr>=1.0.0",
    "datasets>=2.18.0",
    "albumentations>=1.4.0",
]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
]

[project.urls]
Homepage = "https://github.com/jordicatafal/forest-pulse"
Repository = "https://github.com/jordicatafal/forest-pulse"
Issues = "https://github.com/jordicatafal/forest-pulse/issues"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
